+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- Decoupled `fetch-compute-region-disks` step from kms keys dependencies. This
+  allows the Region Disks step to execute successfully even if kms keys steps
+  are disabled.
+
 ## 2.9.0 - 2022-03-09
 
 ### Added