+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to
   allows the Region Disks step to execute successfully even if kms keys steps
   are disabled.
 
+- Decoupled `fetch-memcache-instances` step from compute network dependencies.
+  This allows the Memcache Instances step to execute successfully even if
+  compute network steps are disabled.
+
 ## 2.9.0 - 2022-03-09
 
 ### Added