+++ CHANGELOG.md
@@ -42,6 +42,10 @@ and this project adheres to
   dependencies. This allows the CA Certificate Authorities step to execute
   successfully even if storage steps are disabled.
 
+- Decoupled `fetch-compute-images` step from kms keys dependencies. This allows
+  the Compute Images step to execute successfully even if kms keys steps are
+  disabled.
+
 ## 2.9.0 - 2022-03-09
 
 ### Added