+++ CHANGELOG.md
@@ -6,9 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## 2.9.1 - 2022-03-18
 
-### Changed
+### Fixed
 
 - Decoupled `fetch-compute-region-disks` step from kms keys dependencies. This
   allows the Region Disks step to execute successfully even if kms keys steps